requirements.txt
selenium>=4.17.2
webdriver-manager>=4.0.2
requests>=2.32.3
pypdf>=5.0.0
pdf2image>=1.17.0
Pillow>=10.4.0
pytesseract>=0.3.13
convertdate>=2.4.0
hebrew-tokenizer>=2.3.0
tqdm>=4.66.5
